+++ src/Curiosity.Archiver/IArchiver.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -23,16 +24,7 @@ namespace Curiosity.Archiver
             bool useZip64 = true, 
             CancellationToken cts = default);
 
-        /// <summary>
-        /// Archives specified files in async manner.
-        /// </summary>
-        /// <param name="sourceFiles">Full path to files that will be archive</param>
-        /// <param name="useZip64">Whether use Zip64 for large files support or not. Enabling it may lead to
-        /// incompatibility with old software such as Windows XP and Android prior to 6.0</param>
-        /// <param name="zipFileName">Target zip file name. If <see langword="null"/>, unique name will be generated</param>
-        /// <param name="zipFileNames">Names of files in zip archive. If not <see langword="null"/>, item's count must be equal to <paramref name="sourceFiles"/></param>
-        /// <param name="cts">Cancellation token</param>
-        /// <returns>File stream with zip file in temp directory</returns>
+        [Obsolete("Use method with file names collection")]
         Task<TempFileStream> ZipFilesToStreamAsync(
             IList<string> sourceFiles,
             bool useZip64 = true, 
@@ -43,18 +35,39 @@ namespace Curiosity.Archiver
         /// <summary>
         /// Archives specified files in async manner.
         /// </summary>
-        /// <param name="sourceFiles">Full path to files that will be archive</param>
+        /// <param name="sourceFiles">Collection of the full path to files that will be archived and users file names</param>
         /// <param name="useZip64">Whether use Zip64 for large files support or not. Enabling it may lead to
         /// incompatibility with old software such as Windows XP and Android prior to 6.0</param>
         /// <param name="zipFileName">Target zip file name. If <see langword="null"/>, unique name will be generated</param>
-        /// <param name="zipFileNames">Names of files in zip archive. If not <see langword="null"/>, item's count must be equal to <paramref name="sourceFiles"/></param>
         /// <param name="cts">Cancellation token</param>
-        /// <returns>Full path to zip archive</returns>
+        /// <returns>File stream with zip file in temp directory</returns>
+        Task<TempFileStream> ZipFilesToStreamAsync(
+            IReadOnlyList<FileNames> sourceFiles,
+            bool useZip64 = true, 
+            string? zipFileName = null,
+            CancellationToken cts = default);
+        
+        [Obsolete("Use method with file names collection")]
         Task<string> ZipFilesToFileAsync(
             IList<string> sourceFiles,
             bool useZip64 = true,
             string? zipFileName = null,
             IList<string>? zipFileNames = null,
+            CancellationToken cts = default);
+        
+        /// <summary>
+        /// Archives specified files in async manner.
+        /// </summary>
+        /// <param name="sourceFiles">Collection of the full path to files that will be archived and users file names</param>
+        /// <param name="useZip64">Whether use Zip64 for large files support or not. Enabling it may lead to
+        /// incompatibility with old software such as Windows XP and Android prior to 6.0</param>
+        /// <param name="zipFileName">Target zip file name. If <see langword="null"/>, unique name will be generated</param>
+        /// <param name="cts">Cancellation token</param>
+        /// <returns>Full path to zip archive</returns>
+        Task<string> ZipFilesToFileAsync(
+            IReadOnlyList<FileNames> sourceFiles,
+            bool useZip64 = true,
+            string? zipFileName = null,
             CancellationToken cts = default);
     }
 }